+++ CHANGELOG.md
@@ -55,6 +55,7 @@
 * Replace `moment` with `dayjs`. Refs UIOR-1399.
 * Use a pure function for a template value building. Refs UIOR-1490.
 * Improve fund distribution loading error message. Refs UIOR-1488.
+* Refresh order lines after order mutate actions. Refs UIOR-1499.
 * Use ACQ holdings analyzer to check holdings abandonment. Refs UIOR-1466, UIOR-1468.
 
 ## [8.0.5](https://github.com/folio-org/ui-orders/tree/v8.0.5) (2025-06-30)